[ci] deflake test_actor_cancel test
