+++ .buildkite/macos/macos.rayci.yml
@@ -2,28 +2,28 @@ group: macos tests
 sort_key: "~macos"
 steps:
   # block on premerge
-  - block: "run macos tests"
-    if: build.env("BUILDKITE_PIPELINE_ID") == "0189942e-0876-4b8f-80a4-617f988ec59b"
+  # - block: "run macos tests"
+  #   if: build.env("BUILDKITE_PIPELINE_ID") == "0189942e-0876-4b8f-80a4-617f988ec59b"
 
-  - label: ":tapioca: build: :mac: wheels and jars (x86_64)"
-    if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
-    tags:
-      - macos_wheels
-      - python_dependencies
-    job_env: MACOS
-    instance_type: macos
-    commands:
-      - ./ci/ray_ci/macos/macos_ci_build.sh build_x86_64
+  # - label: ":tapioca: build: :mac: wheels and jars (x86_64)"
+  #   if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
+  #   tags:
+  #     - macos_wheels
+  #     - python_dependencies
+  #   job_env: MACOS
+  #   instance_type: macos
+  #   commands:
+  #     - ./ci/ray_ci/macos/macos_ci_build.sh build_x86_64
 
-  - label: ":tapioca: build: :mac: wheels and jars (aarch64)"
-    if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
-    tags:
-      - macos_wheels
-      - python_dependencies
-    job_env: MACOS
-    instance_type: macos-arm64
-    commands:
-      - ./ci/ray_ci/macos/macos_ci_build.sh build_aarch64
+  # - label: ":tapioca: build: :mac: wheels and jars (aarch64)"
+  #   if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
+  #   tags:
+  #     - macos_wheels
+  #     - python_dependencies
+  #   job_env: MACOS
+  #   instance_type: macos-arm64
+  #   commands:
+  #     - ./ci/ray_ci/macos/macos_ci_build.sh build_aarch64
 
   - label: ":ray: core: :mac: small & client tests"
     if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
@@ -37,94 +37,94 @@ steps:
     commands:
       - ./ci/ray_ci/macos/macos_ci.sh run_small_test
 
-  - label: ":ray: core: :mac: medium a-j tests"
-    if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
-    tags: 
-      - core_cpp
-      - python
-      - macos_wheels
-      - oss
-    job_env: MACOS
-    instance_type: macos
-    commands:
-      - ./ci/ray_ci/macos/macos_ci.sh run_medium_a_j_test
+  # - label: ":ray: core: :mac: medium a-j tests"
+  #   if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
+  #   tags: 
+  #     - core_cpp
+  #     - python
+  #     - macos_wheels
+  #     - oss
+  #   job_env: MACOS
+  #   instance_type: macos
+  #   commands:
+  #     - ./ci/ray_ci/macos/macos_ci.sh run_medium_a_j_test
 
-  - label: ":ray: core: :mac: medium k-z tests"
-    if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
-    tags: 
-      - core_cpp
-      - python
-      - macos_wheels
-      - oss
-    job_env: MACOS
-    instance_type: macos
-    commands:
-      - ./ci/ray_ci/macos/macos_ci.sh run_medium_k_z_test
+  # - label: ":ray: core: :mac: medium k-z tests"
+  #   if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
+  #   tags: 
+  #     - core_cpp
+  #     - python
+  #     - macos_wheels
+  #     - oss
+  #   job_env: MACOS
+  #   instance_type: macos
+  #   commands:
+  #     - ./ci/ray_ci/macos/macos_ci.sh run_medium_k_z_test
 
-  - label: ":ray: core: :mac: large tests"
-    if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
-    tags: 
-      - core_cpp
-      - python
-      - macos_wheels
-      - oss
-    job_env: MACOS
-    instance_type: macos
-    parallelism: 3
-    commands:
-      - ./ci/ray_ci/macos/macos_ci.sh run_large_test
+  # - label: ":ray: core: :mac: large tests"
+  #   if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
+  #   tags: 
+  #     - core_cpp
+  #     - python
+  #     - macos_wheels
+  #     - oss
+  #   job_env: MACOS
+  #   instance_type: macos
+  #   parallelism: 3
+  #   commands:
+  #     - ./ci/ray_ci/macos/macos_ci.sh run_large_test
 
-  - label: ":ray: core: :mac: core and dashboard tests"
-    if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
-    tags:
-      - core_cpp
-      - python
-      - dashboard
-      - macos_wheels
-      - oss
-    job_env: MACOS
-    instance_type: macos
-    soft_fail: true
-    commands:
-      - ./ci/ray_ci/macos/macos_ci.sh run_core_dashboard_test
+  # - label: ":ray: core: :mac: core and dashboard tests"
+  #   if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
+  #   tags:
+  #     - core_cpp
+  #     - python
+  #     - dashboard
+  #     - macos_wheels
+  #     - oss
+  #   job_env: MACOS
+  #   instance_type: macos
+  #   soft_fail: true
+  #   commands:
+  #     - ./ci/ray_ci/macos/macos_ci.sh run_core_dashboard_test
 
-  - label: ":ray: core: :mac: core c++ and java tests"
-    if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
-    tags:
-      - cpp
-      - java
-      - macos_wheels
-      - oss
-    job_env: MACOS
-    instance_type: macos
-    soft_fail: true
-    commands:
-      - RAY_INSTALL_JAVA=1 ./ci/ray_ci/macos/macos_ci.sh run_ray_cpp_and_java
+  # - label: ":ray: core: :mac: core c++ and java tests"
+  #   if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
+  #   tags:
+  #     - cpp
+  #     - java
+  #     - macos_wheels
+  #     - oss
+  #   job_env: MACOS
+  #   instance_type: macos
+  #   soft_fail: true
+  #   commands:
+  #     - RAY_INSTALL_JAVA=1 ./ci/ray_ci/macos/macos_ci.sh run_ray_cpp_and_java
 
-  - label: ":ray: core: :mac: small & large flaky tests"
-    if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
-    tags: 
-      - core_cpp
-      - python
-      - macos_wheels
-      - oss
-      - skip_on_premerge
-    job_env: MACOS
-    instance_type: macos
-    soft_fail: true
-    commands:
-      - ./ci/ray_ci/macos/macos_ci.sh run_small_and_large_flaky_tests
+  # - label: ":ray: core: :mac: small & large flaky tests"
+  #   if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
+  #   tags: 
+  #     - core_cpp
+  #     - python
+  #     - macos_wheels
+  #     - oss
+  #     - skip_on_premerge
+  #   job_env: MACOS
+  #   instance_type: macos
+  #   soft_fail: true
+  #   commands:
+  #     - ./ci/ray_ci/macos/macos_ci.sh run_small_and_large_flaky_tests
 
-  - label: ":ray: core: :mac: medium flaky tests"
-    if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
-    tags: 
-      - core_cpp
-      - python
-      - macos_wheels
-      - oss
-      - skip_on_premerge
-    job_env: MACOS
-    instance_type: macos
-    soft_fail: true
-    commands:
-      - ./ci/ray_ci/macos/macos_ci.sh run_medium_flaky_tests
+  # - label: ":ray: core: :mac: medium flaky tests"
+  #   if: build.env("BUILDKITE_PIPELINE_ID") != "0189e759-8c96-4302-b6b5-b4274406bf89"
+  #   tags: 
+  #     - core_cpp
+  #     - python
+  #     - macos_wheels
+  #     - oss
+  #     - skip_on_premerge
+  #   job_env: MACOS
+  #   instance_type: macos
+  #   soft_fail: true
+  #   commands:
+  #     - ./ci/ray_ci/macos/macos_ci.sh run_medium_flaky_tests
